parseRtasVpd() - Fix the possibility of overflow
Fix the possibility of overflow, while calculating the
size of every elements (each of the size if 4096). With
int only 8 of them can fit in. Increase it to unsigned
long int 2^31 to avoid overflow.

Also replace all the places where the result from parseRtasVpd(),
is used.

Cc: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>
[Fixed compilation warning - Vasant]

+++ src/include/devicetreecollector.hpp
@@ -278,7 +278,7 @@ namespace lsvpd
 			 *   The size of the rtasData buffer.
 			 */
 			void parseRtasVpd( vector<Component*>& devs, char *rtasData,
-				int rtasDataSize);
+				unsigned long int rtasDataSize);
 
 			/**
 			 * Pares the System VPD  and add the information to the
